Convert FormSelect to ts, add types for label. add support for hiding with prop
